+++ client/web/src/enterprise/code-monitoring/CodeMonitoringNode.tsx
@@ -1,7 +1,7 @@
 import * as H from 'history'
 import React, { useState, useCallback, useMemo } from 'react'
 import { Observable, concat, of } from 'rxjs'
-import { switchMap, catchError, startWith, takeUntil, tap, delay, mergeMap } from 'rxjs/operators'
+import { switchMap, catchError, startWith, takeUntil, tap, delay } from 'rxjs/operators'
 
 import { Toggle } from '@sourcegraph/branded/src/components/Toggle'
 import { ErrorLike, isErrorLike, asError } from '@sourcegraph/common'
@@ -9,14 +9,12 @@ import { Button, LoadingSpinner, useEventObservable, Link, Alert } from '@source
 
 import { CodeMonitorFields, ToggleCodeMonitorEnabledResult } from '../../graphql-operations'
 
-import { sendTestEmail, toggleCodeMonitorEnabled as _toggleCodeMonitorEnabled } from './backend'
+import { toggleCodeMonitorEnabled as _toggleCodeMonitorEnabled } from './backend'
 import styles from './CodeMonitoringNode.module.scss'
 
 export interface CodeMonitorNodeProps {
     node: CodeMonitorFields
     location: H.Location
-    isSiteAdminUser: boolean
-    showCodeMonitoringTestEmailButton: boolean
 
     toggleCodeMonitorEnabled?: typeof _toggleCodeMonitorEnabled
 }
@@ -26,8 +24,6 @@ const LOADING = 'LOADING' as const
 export const CodeMonitorNode: React.FunctionComponent<CodeMonitorNodeProps> = ({
     location,
     node,
-    isSiteAdminUser,
-    showCodeMonitoringTestEmailButton,
     toggleCodeMonitorEnabled = _toggleCodeMonitorEnabled,
 }: CodeMonitorNodeProps) => {
     const [enabled, setEnabled] = useState<boolean>(node.enabled)
@@ -62,22 +58,23 @@ export const CodeMonitorNode: React.FunctionComponent<CodeMonitorNodeProps> = ({
         )
     )
 
-    const [sendEmailRequest] = useEventObservable(
-        useCallback(
-            (click: Observable<React.MouseEvent<HTMLButtonElement>>) =>
-                click.pipe(
-                    mergeMap(() =>
-                        sendTestEmail(node.trigger.id).pipe(
-                            startWith(LOADING),
-                            catchError(error => [asError(error)])
-                        )
-                    )
-                ),
-            [node]
-        )
+    const actions = useMemo(
+        () =>
+            node.actions.nodes
+                .map(action => {
+                    switch (action.__typename) {
+                        case 'MonitorEmail':
+                            return 'Sends email notification'
+                        case 'MonitorSlackWebhook':
+                            return 'Sends Slack notification'
+                        default:
+                            return ''
+                    }
+                })
+                .filter(string => string !== '')
+                .join('; '),
+        [node.actions]
     )
-
-    const hasEnabledAction = useMemo(() => node.actions.nodes.filter(node => node.enabled).length > 0, [node.actions])
 
     return (
         <div className={styles.codeMonitoringNode}>
@@ -86,16 +83,8 @@ export const CodeMonitorNode: React.FunctionComponent<CodeMonitorNodeProps> = ({
                     <div className="font-weight-bold">
                         <Link to={`${location.pathname}/${node.id}`}>{node.description}</Link>
                     </div>
-                    {/** TODO: Generate this text based on the type of action when new actions are added. */}
                     {node.actions.nodes.length > 0 && (
-                        <div className="d-flex text-muted">
-                            New search result → Sends email notifications{' '}
-                            {showCodeMonitoringTestEmailButton && isSiteAdminUser && hasEnabledAction && node.enabled && (
-                                <Button className="p-0 border-0 ml-2" onClick={sendEmailRequest} variant="link">
-                                    Send test email
-                                </Button>
-                            )}
-                        </div>
+                        <div className="d-flex text-muted align-items-center">New search result → {actions}</div>
                     )}
                 </div>
                 <div className="d-flex">