+++ client/web/src/enterprise/code-monitoring/components/actions/ActionEditor.tsx
@@ -6,38 +6,48 @@ import { Button, Card } from '@sourcegraph/wildcard'
 
 import styles from '../CodeMonitorForm.module.scss'
 
-interface Props {
+export interface ActionEditorProps {
     title: React.ReactNode
+    label: string // Similar to title, but for string-only labels
     subtitle: string
     disabled: boolean
     completed: boolean
     completedSubtitle: string
+    idName: string // Name used for generating IDs, including form control IDs and test IDs
 
     actionEnabled: boolean
     toggleActionEnabled: (enabled: boolean) => void
 
+    canSubmit?: boolean
     onSubmit: React.FormEventHandler
     onCancel?: React.FormEventHandler
 
     canDelete: boolean
     onDelete: React.FormEventHandler
+
+    // For testing purposes only
+    _testStartOpen?: boolean
 }
 
-export const ActionEditor: React.FunctionComponent<Props> = ({
+export const ActionEditor: React.FunctionComponent<ActionEditorProps> = ({
     title,
+    label,
     subtitle,
     disabled,
     completed,
     completedSubtitle,
+    idName,
     actionEnabled,
     toggleActionEnabled,
+    canSubmit = true,
     onSubmit,
     onCancel,
     canDelete,
     onDelete,
     children,
+    _testStartOpen = false,
 }) => {
-    const [expanded, setExpanded] = useState(false)
+    const [expanded, setExpanded] = useState(_testStartOpen)
     const toggleExpanded = useCallback(() => {
         if (!disabled) {
             setExpanded(expanded => !expanded)
@@ -70,10 +80,6 @@ export const ActionEditor: React.FunctionComponent<Props> = ({
         [onDelete]
     )
 
-    // When the action is completed, the wrapper cannot be a button because we show nested buttons inside it.
-    // Use a div instead. The edit button will still allow keyboard users to activate the form.
-    const CollapsedWrapperElement = completed ? 'div' : Button
-
     return (
         <>
             {expanded && (
@@ -90,31 +96,41 @@ export const ActionEditor: React.FunctionComponent<Props> = ({
                                 value={actionEnabled}
                                 onToggle={toggleActionEnabled}
                                 className="mr-2"
-                                aria-labelledby="code-monitoring-form-actions-enable-toggle"
+                                aria-labelledby={`code-monitoring-${idName}-form-actions-enable-toggle`}
+                                data-testid={`enable-action-toggle-expanded-${idName}`}
                             />
                         </div>
-                        <span id="code-monitoring-form-actions-enable-toggle">
+                        <span id={`code-monitoring-${idName}-form-actions-enable-toggle`}>
                             {actionEnabled ? 'Enabled' : 'Disabled'}
                         </span>
                     </div>
                     <div className="d-flex justify-content-between">
                         <div>
                             <Button
-                                type="submit"
-                                data-testid="submit-action"
-                                className="mr-1 test-submit-action"
+                                data-testid={`submit-action-${idName}`}
+                                className={`mr-1 test-submit-action-${idName}`}
                                 onClick={submitHandler}
-                                onSubmit={submitHandler}
+                                disabled={!canSubmit}
                                 variant="secondary"
                             >
                                 Continue
                             </Button>
-                            <Button onClick={cancelHandler} outline={true} variant="secondary">
+                            <Button
+                                onClick={cancelHandler}
+                                outline={true}
+                                variant="secondary"
+                                data-testid={`cancel-action-${idName}`}
+                            >
                                 Cancel
                             </Button>
                         </div>
                         {canDelete && (
-                            <Button onClick={deleteHandler} outline={true} variant="danger">
+                            <Button
+                                onClick={deleteHandler}
+                                outline={true}
+                                variant="danger"
+                                data-testid={`delete-action-${idName}`}
+                            >
                                 Delete
                             </Button>
                         )}
@@ -123,11 +139,13 @@ export const ActionEditor: React.FunctionComponent<Props> = ({
             )}
             {!expanded && (
                 <Card
-                    as={CollapsedWrapperElement}
-                    data-testid="form-action-toggle-email-notification"
-                    className={classNames('test-action-button', styles.cardButton, disabled && 'disabled')}
+                    // When the action is completed, the wrapper cannot be a button because we show nested buttons inside it.
+                    // Use a div instead. The edit button will still allow keyboard users to activate the form.
+                    as={completed ? 'div' : Button}
+                    data-testid={`form-action-toggle-${idName}`}
+                    className={classNames(`test-action-button-${idName}`, styles.cardButton, disabled && 'disabled')}
                     disabled={disabled}
-                    aria-label="Edit action: Send email notifications"
+                    aria-label={`Edit action: ${label}`}
                     onClick={toggleExpanded}
                 >
                     <div className="d-flex justify-content-between align-items-center w-100">
@@ -141,7 +159,10 @@ export const ActionEditor: React.FunctionComponent<Props> = ({
                                 {title}
                             </div>
                             {completed ? (
-                                <span className="text-muted font-weight-normal" data-testid="existing-action-email">
+                                <span
+                                    className="text-muted font-weight-normal"
+                                    data-testid={`existing-action-${idName}`}
+                                >
                                     {completedSubtitle}
                                 </span>
                             ) : (
@@ -156,6 +177,7 @@ export const ActionEditor: React.FunctionComponent<Props> = ({
                                         value={actionEnabled}
                                         onToggle={toggleActionEnabled}
                                         className="mr-3"
+                                        data-testid={`enable-action-toggle-collapsed-${idName}`}
                                     />
                                 </div>
                                 <Button variant="link" className="p-0">